# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Bissi / Committee / Chit Fund Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /loans:
    get:
      operationId: listLoans
      tags: [loans]
      summary: List loans
      parameters:
        - name: customerId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
        - name: branchId
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of loans
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoanListResponse"
    post:
      operationId: createLoan
      tags: [loans]
      summary: Create/apply for a loan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoanInput"
      responses:
        "201":
          description: Loan created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
  /loans/{id}:
    get:
      operationId: getLoan
      tags: [loans]
      summary: Get loan by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Loan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
    patch:
      operationId: updateLoan
      tags: [loans]
      summary: Update loan status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoanUpdate"
      responses:
        "200":
          description: Loan updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
  /loans/{id}/emi-schedule:
    get:
      operationId: getLoanEmiSchedule
      tags: [loans]
      summary: Get EMI schedule for a loan
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: EMI schedule
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmiEntry"
...
  /lotteries:
    get:
      operationId: listLotteries
      tags: [lotteries]
      summary: List lotteries
      parameters:
        - name: committeeId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of lotteries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lottery"
    post:
      operationId: createLottery
      tags: [lotteries]
      summary: Create a lottery draw
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LotteryInput"
      responses:
        "201":
          description: Lottery created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lottery"
  /lotteries/{id}:
    get:
      operationId: getLottery
      tags: [lotteries]
      summary: Get lottery by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lottery
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lottery"
    patch:
      operationId: updateLottery
      tags: [lotteries]
      summary: Update lottery / declare winner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LotteryUpdate"
      responses:
        "200":
          description: Lottery updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lottery"
  /lotteries/{id}/draw:
    post:
      operationId: conductDraw
      tags: [lotteries]
      summary: Conduct a lottery draw (auto-select winner)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Draw result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lottery"

  # Reports
...
components:
  schemas:
...
    Loan:
      type: object
      required: [id, customerId, principalAmount, interestRate, interestType, tenure, status, branchId, createdAt]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: ["string", "null"]
        customerMobile:
          type: ["string", "null"]
        principalAmount:
          type: number
        interestRate:
          type: number
        interestType:
          type: string
          enum: [flat, reducing]
        tenure:
          type: integer
        emiAmount:
          type: ["number", "null"]
        totalAmount:
          type: ["number", "null"]
        paidAmount:
          type: number
        outstandingAmount:
          type: ["number", "null"]
        status:
          type: string
          enum: [pending, approved, active, closed, rejected, overdue]
        branchId:
          type: integer
        branchName:
          type: ["string", "null"]
        disbursedAt:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        createdAt:
          type: string
    LoanInput:
      type: object
      required: [customerId, principalAmount, interestRate, interestType, tenure, branchId]
      properties:
        customerId:
          type: integer
        principalAmount:
          type: number
        interestRate:
          type: number
        interestType:
          type: string
          enum: [flat, reducing]
        tenure:
          type: integer
        branchId:
          type: integer
        purpose:
          type: string
    LoanUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, approved, active, closed, rejected, overdue]
        disbursedAt:
          type: string
    LoanListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Loan"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    EmiEntry:
      type: object
      required: [installmentNumber, dueDate, emiAmount, principalComponent, interestComponent, status]
      properties:
        installmentNumber:
          type: integer
        dueDate:
          type: string
        emiAmount:
          type: number
        principalComponent:
          type: number
        interestComponent:
          type: number
        paidAmount:
          type: number
        status:
          type: string
          enum: [pending, paid, overdue, partial]
...
    Lottery:
      type: object
      required: [id, committeeId, drawDate, status, createdAt]
      properties:
        id:
          type: integer
        committeeId:
          type: integer
        committeeName:
          type: ["string", "null"]
        drawDate:
          type: string
        winnerId:
          type: ["integer", "null"]
        winnerName:
          type: ["string", "null"]
        winnerToken:
          type: ["string", "null"]
        prizeAmount:
          type: ["number", "null"]
        status:
          type: string
          enum: [scheduled, completed, cancelled]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    LotteryInput:
      type: object
      required: [committeeId, drawDate]
      properties:
        committeeId:
          type: integer
        drawDate:
          type: string
        prizeAmount:
          type: number
        notes:
          type: string
    LotteryUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [scheduled, completed, cancelled]
        winnerId:
          type: integer
        prizeAmount:
          type: number
        notes:
          type: string
...